+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### 1.34
 
+  * Bugfix:  #292 Account for metadata size during size chunking in ChunkStore
   * Feature: #283 Support for all pandas frequency strings in ChunkStore DateChunker
   * Feature: #286 Add has_symbol to ChunkStore and support for partial symbol matching in list_symbols